requirements.txt
Flask==2.0.1
requests==2.26.0
python-dotenv==0.19.1
Flask-Cors==3.0.10
ffmpeg-python==0.2.0